Add a build option to disable malloc poisoning in the simple allocator
This change:

* Adds a Meson option, 'disable-malloc-poison=<bool>',
* Translates that option into a CPP define, DISABLE_MALLOC_POISON, and
* Disables the simple allocator's poisoning behavior

This change is useful for cheesecloth example programs where malloc
poisoning is not needed or desirable.

This change only affects the simple allocator's malloc() implementation
(i.e. when FROMAGER_SIMPLE_MALLOC is set). The simple allocator's free()
implementation does not implement poisoning at this time, so this commit
just updates the TODO comment there so that a future update can respect
DISABLE_MALLOC_POISON.